+++ types/typings-test.ts
@@ -16,6 +16,8 @@ const optionsFull: IDatePickerOptions = {
   locale: 'en_GB',
   okText: 'Select',
   cancelText: 'Cancel',
+  clearText: 'Clear',
+  titleText: 'Picker title',
   android: {
     theme: 5, // Theme_DeviceDefault_Dialog
     calendar: false,